Add a --dev option to run_docker.py
This is meant for modifying AlphaFold without re-building
the Docker image.

A Dockerfile for an alphafold-dev image is defined.
If --dev is specified, run_docker.py creates an alphafold-dev container
by default.
The AlphaFold source code is bound to this container
by mounting an external location, defined by $ALPHAFOLD_DIR